Refactor: 하루섭취 자연식 컨트롤러를 자연식 컨트롤러로 병합 (#46)
